feature(sign up): abstract sign up form into a component
